+++ cloudamqp/resource_cloudamqp_plugin.go
@@ -47,34 +47,70 @@ func resourcePlugin() *schema.Resource {
 				Computed:    true,
 				Description: "The version of the plugin",
 			},
+			"sleep": {
+				Type:        schema.TypeInt,
+				Optional:    true,
+				Default:     10,
+				Description: "Configurable sleep time in seconds between retries for plugins",
+			},
+			"timeout": {
+				Type:        schema.TypeInt,
+				Optional:    true,
+				Default:     1800,
+				Description: "Configurable timeout time in seconds for plugins",
+			},
 		},
 	}
 }
 
 func resourcePluginCreate(d *schema.ResourceData, meta interface{}) error {
-	api := meta.(*api.API)
-	_, err := api.EnablePlugin(d.Get("instance_id").(int), d.Get("name").(string))
+	var (
+		api        = meta.(*api.API)
+		instanceID = d.Get("instance_id").(int)
+		name       = d.Get("name").(string)
+		sleep      = d.Get("sleep").(int)
+		timeout    = d.Get("timeout").(int)
+	)
+
+	log.Printf("[DEBUG] create plugin instanceID: %v, name: %v, sleep: %v, timeout: %v",
+		instanceID, name, sleep, timeout)
+	_, err := api.EnablePlugin(instanceID, name, sleep, timeout)
 	if err != nil {
 		return err
 	}
-	d.SetId(d.Get("name").(string))
+	d.SetId(name)
 	return resourcePluginRead(d, meta)
 }
 
 func resourcePluginRead(d *schema.ResourceData, meta interface{}) error {
+	var (
+		api        = meta.(*api.API)
+		instanceID = d.Get("instance_id").(int)
+		name       = d.Get("name").(string)
+		sleep      = d.Get("sleep").(int)
+		timeout    = d.Get("timeout").(int)
+	)
+
+	// Support for importing resource
 	if strings.Contains(d.Id(), ",") {
+		log.Printf("[DEBUG] import plugin instanceID: %v, id: %v", instanceID, d.Id())
 		s := strings.Split(d.Id(), ",")
-		d.SetId(s[0])
-		d.Set("name", s[0])
-		instanceID, _ := strconv.Atoi(s[1])
+		name = s[0]
+		d.SetId(name)
+		d.Set("name", name)
+		instanceID, _ = strconv.Atoi(s[1])
 		d.Set("instance_id", instanceID)
+		// Set default values for optional arguments
+		d.Set("sleep", 10)
+		d.Set("timeout", 1800)
 	}
-	if d.Get("instance_id").(int) == 0 {
+	if instanceID == 0 {
 		return errors.New("missing instance identifier: {resource_id},{instance_id}")
 	}
 
-	api := meta.(*api.API)
-	data, err := api.ReadPlugin(d.Get("instance_id").(int), d.Get("name").(string))
+	log.Printf("[DEBUG] import plugin instanceID: %v, name: %v, sleep: %v, timeout: %v",
+		instanceID, name, sleep, timeout)
+	data, err := api.ReadPlugin(instanceID, name, sleep, timeout)
 	if err != nil {
 		return err
 	}
@@ -91,16 +127,16 @@ func resourcePluginRead(d *schema.ResourceData, meta interface{}) error {
 }
 
 func resourcePluginUpdate(d *schema.ResourceData, meta interface{}) error {
-	api := meta.(*api.API)
-	keys := []string{"name", "enabled"}
-	params := make(map[string]interface{})
-	for _, k := range keys {
-		if v := d.Get(k); v != nil {
-			params[k] = v
-		}
-	}
+	var (
+		api        = meta.(*api.API)
+		instanceID = d.Get("instance_id").(int)
+		name       = d.Get("name").(string)
+		enabled    = d.Get("enabled").(bool)
+		sleep      = d.Get("sleep").(int)
+		timeout    = d.Get("timeout").(int)
+	)
 
-	_, err := api.UpdatePlugin(d.Get("instance_id").(int), params)
+	_, err := api.UpdatePlugin(instanceID, name, enabled, sleep, timeout)
 	if err != nil {
 		return fmt.Errorf("[Failed to update pluign: %v", err)
 	}
@@ -108,13 +144,20 @@ func resourcePluginUpdate(d *schema.ResourceData, meta interface{}) error {
 }
 
 func resourcePluginDelete(d *schema.ResourceData, meta interface{}) error {
+	var (
+		api        = meta.(*api.API)
+		instanceID = d.Get("instance_id").(int)
+		name       = d.Get("name").(string)
+		sleep      = d.Get("sleep").(int)
+		timeout    = d.Get("timeout").(int)
+	)
+
 	if enableFasterInstanceDestroy {
 		log.Printf("[DEBUG] cloudamqp::resource::plugin::delete skip calling backend.")
 		return nil
 	}
 
-	api := meta.(*api.API)
-	return api.DeletePlugin(d.Get("instance_id").(int), d.Get("name").(string))
+	return api.DeletePlugin(instanceID, name, sleep, timeout)
 }
 
 func validatePluginSchemaAttribute(key string) bool {